+++ src/components/intro/index.tsx
@@ -1,0 +1,6 @@
+/* 
+import 정리용 페이지
+*/
+
+export { IntroService } from './IntroService';
+export { IntroTeam } from './IntroTeam';
