+++ src/components/intro/introService/index.tsx
@@ -1,0 +1,7 @@
+/* 
+import 정리용 페이지
+*/
+
+export { IntroductionPart } from './IntroductionPart';
+export { DescribeMainPart } from './DescribeMainPart';
+export { DescribeSubPart } from './DescribeSubPart';
